+++ api/app/api.py
@@ -4,9 +4,8 @@ from fastapi import FastAPI, HTTPException
 from fastapi.middleware.cors import CORSMiddleware
 import pandas as pd
 from dotenv import load_dotenv
-from .schemas import ForecastRequest, TokenRequest
 from quartz_solar_forecast.forecast import run_forecast
-from quartz_solar_forecast.pydantic_models import PVSite
+from quartz_solar_forecast.pydantic_models import PVSite, ForecastRequest, TokenRequest
 from .utils.enphase_utils import get_enphase_auth_url, get_enphase_access_token, get_enphase_data, run_forecast_for_enphase
 
 load_dotenv()
